+++ app/views/play/cinematic/Command/CinematicParser.js
@@ -82,10 +82,10 @@ const parseDialogNode = (dialogNode, systems) =>
  * @param {Object} systems The systems.
  * @returns {AbstractCommand[][]} A 2d array of commands.
  */
-export const parseShot = (shot, systems) => {
-  const setupCommands = parseSetup(shot, systems) || []
+export const parseShot = (shot, cinematicSystems) => {
+  const setupCommands = parseSetup(shot, cinematicSystems) || []
   const dialogNodes = (shot.dialogNodes || [])
-    .map((node) => parseDialogNode(node, systems))
+    .map((node) => parseDialogNode(node, cinematicSystems))
     .filter(dialogCommands => dialogCommands.length > 0)
 
   // If we have both dialogNodes and some setupCommands we want to
