+++ app/views/play/cinematic/cinematicController.js
@@ -74,8 +74,6 @@ export class CinematicController {
 
     this.commands = []
 
-    this.commands = []
-
     this.startUp()
   }
